+++ src/applications/vaos/services/mocks/utils/confirmedAppointments.js
@@ -370,23 +370,37 @@ const appointmentTemplates = {
           status: 'booked',
           patientIcn: null,
           modality: 'communityCareEps',
-          created: formatAppointmentDate(new Date()),
-          locationId: 'sandbox-network-5vuTac8v',
-          clinic: '9L5WE4aN',
-          start: formatAppointmentDate(startDate),
-          referralId: 'VA0000007254',
-          referral: {
-            referralNumber: 'VA0000007254',
-          },
-          providerServiceId: '9L5WE4aN',
-          providerName: 'A&D HEALTH CARE PROFESSIONALS, INC.',
-          location: 'Error fetching facility details',
           kind: 'cc',
           type: 'COMMUNITY_CARE_APPOINTMENT',
-          pending: false,
+          start: formatAppointmentDate(startDate),
           past: startDate < new Date(),
           future: startDate > new Date(),
+          lastRetrieved: new Date().toISOString(),
+          isLatest: true,
           cancellable: false, // CC appointments that are booked are not cancellable
+          provider: {
+            id: appointmentId,
+            name: 'Dr. Smith @ Acme Cardiology - Anywhere, USA',
+            practice: 'Acme Cardiology',
+            phone: '555-555-0001',
+            location: {
+              name: 'Temple University Hospital - Jeanes Campus',
+              address: '7500 CENTRAL AVE, STE 108, PHILADELPHIA, PA 19111-2430',
+              latitude: 40.06999282694126,
+              longitude: -75.08769957031448,
+              timezone: 'America/New_York',
+            },
+          },
+          location: {
+            id: appointmentId,
+            type: 'appointments',
+            attributes: {
+              name: 'Temple University Hospital - Jeanes Campus',
+              timezone: {
+                timeZoneId: 'America/New_York',
+              },
+            },
+          },
         },
       };
     },
